requirements.txt
aiohttp==3.8.2
aiosignal==1.3.1
alpaca-trade-api==3.0.2
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
deprecation==2.1.0
executing==2.0.1
fonttools==4.45.0
frozendict==2.3.9
frozenlist==1.4.0
html5lib==1.1
idna==3.4
ipykernel==6.27.0
ipython==8.17.2
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
msgpack==1.0.3
multidict==5.2.0
multitasking==0.0.11
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
peewee==3.17.0
pexpect==4.8.0
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.13.0
tzdata==2023.3
urllib3==1.26.18
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==1.6.4
websockets==10.4
yarl==1.9.2
yfinance==0.2.32
